MINIFICPP-1373 - Move ConsumeKafkaMaxPollTimeValidator to ConsumeKafka.cpp
